% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_suffix}
\alias{add_suffix}
\title{helper function to switch labels between english and german for specific variables}
\usage{
add_suffix(strings, language = "de")
}
\description{
helper function to switch labels between english and german for specific variables
}
\keyword{internal}
