% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Application Server}
\usage{
app_server(input, output, session)
}
\description{
Define the server logic for the SUF-Explorer Shiny application.
Manages reactive modules, dataset exploration, data transformation, and sidebar updates.
}
\keyword{internal}
