% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize_first_letter}
\alias{capitalize_first_letter}
\title{Helper function to capitalize the first letter of a string}
\usage{
capitalize_first_letter(x)
}
\description{
Helper function to capitalize the first letter of a string
}
\keyword{internal}
