% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dataframe}
\alias{create_dataframe}
\title{Create dataframe from dataset and variable names}
\usage{
create_dataframe(dataset, variable)
}
\description{
Create dataframe from dataset and variable names
}
\keyword{internal}
