% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dataset_names}
\alias{create_dataset_names}
\title{Create dataset names from linkage keys}
\usage{
create_dataset_names(datapath)
}
\description{
Create dataset names from linkage keys
}
\keyword{internal}
