% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_linkage_data}
\alias{create_linkage_data}
\title{Create linkage dataset for merging in data transformation}
\usage{
create_linkage_data(datapath)
}
\description{
Create linkage dataset for merging in data transformation
}
\keyword{internal}
