% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_transformation.R
\name{data_transformation_add_variables_ui}
\alias{data_transformation_add_variables_ui}
\title{Module: Data Transformation}
\usage{
data_transformation_add_variables_ui(id)
}
\description{
This module provides UI and server logic for transforming NEPS SUF data
into a person-year format, including variable selection, spell prioritization,
script preview, and download.
}
\keyword{internal}
