% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_transformation.R
\name{data_transformation_prio_ui}
\alias{data_transformation_prio_ui}
\title{UI func for priorisation tab}
\usage{
data_transformation_prio_ui(id)
}
\description{
UI func for priorisation tab
}
\keyword{internal}
