% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_transformation.R
\name{data_transformation_server}
\alias{data_transformation_server}
\title{func for server}
\usage{
data_transformation_server(id, settings_reactive)
}
\description{
func for server
}
\keyword{internal}
