% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_transformation.R
\name{data_transformation_sidebar_ui}
\alias{data_transformation_sidebar_ui}
\title{UI func for sidebar}
\usage{
data_transformation_sidebar_ui(id)
}
\description{
UI func for sidebar
}
\keyword{internal}
