% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{dataset_exploration_card}
\alias{dataset_exploration_card}
\title{Cards for Dataset Exploration and Data Transformation}
\usage{
dataset_exploration_card()
}
\value{
`bslib::navset_card_tab` objects

A `navset_card_tab` object for Shiny UI.
}
\description{
These functions create UI cards for the Shiny app:
- `dataset_exploration_card()` for exploring datasets
- `cards_data_trans()` for data transformation
}
\keyword{Card}
\keyword{Creates}
\keyword{Dataset}
\keyword{Exploration}
\keyword{Shows}
\keyword{Table:}
\keyword{Variable}
\keyword{`navset_card_tab`}
\keyword{a}
\keyword{containing}
\keyword{data}
\keyword{dataset}
\keyword{dropdown.}
\keyword{exploration:}
\keyword{for}
\keyword{internal}
\keyword{overview}
\keyword{settings}
\keyword{tab}
\keyword{table}
\keyword{with}
