% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_exploration.R
\name{dataset_ui}
\alias{dataset_ui}
\title{Data Exploration Module}
\usage{
dataset_ui(id)
}
\description{
Internal Shiny module for exploring datasets with multiple cohort selection.
}
\keyword{UI}
\keyword{func}
\keyword{internal}
