% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_suf_version}
\alias{extract_suf_version}
\title{extracts the latest suf version from the datafiles name of the suf data in given datapath}
\usage{
extract_suf_version(datapath, short = FALSE)
}
\description{
extracts the latest suf version from the datafiles name of the suf data in given datapath
}
\keyword{internal}
