% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_dataframes}
\alias{filter_dataframes}
\title{Filter a list of dataframes by variable names}
\usage{
filter_dataframes(list_of_dfs, vars)
}
\description{
Filter a list of dataframes by variable names
}
\keyword{internal}
