% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_var}
\alias{filter_var}
\title{function that filters a dataset depending on variable type}
\usage{
filter_var(x, val)
}
\description{
function that filters a dataset depending on variable type
}
\keyword{internal}
