% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_functions.R
\name{further_training_gen_r}
\alias{further_training_gen_r}
\title{helper function to put these further education prep lines into multiple script versions. R.}
\usage{
further_training_gen_r(english, suf_version_short)
}
\description{
helper function to put these further education prep lines into multiple script versions. R.
}
\keyword{internal}
