% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_functions.R
\name{further_training_gen_stata}
\alias{further_training_gen_stata}
\title{helper function to put these further education prep lines into multiple script versions. Stata.}
\usage{
further_training_gen_stata(english, suf_version_short)
}
\description{
helper function to put these further education prep lines into multiple script versions. Stata.
}
\keyword{internal}
