% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_functions.R
\name{gen_children_example_stata_sc3_4}
\alias{gen_children_example_stata_sc3_4}
\title{function to generate child example prep. Stata SC3 SC4.}
\usage{
gen_children_example_stata_sc3_4(english, sc, suf_version)
}
\description{
function to generate child example prep. Stata SC3 SC4.
}
\keyword{internal}
