% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_functions.R
\name{gen_children_example_stata_sc5_6}
\alias{gen_children_example_stata_sc5_6}
\title{function to generate child example prep. Stata SC5 SC6.}
\usage{
gen_children_example_stata_sc5_6(english, sc, suf_version)
}
\description{
function to generate child example prep. Stata SC5 SC6.
}
\keyword{internal}
