% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_comb_char}
\alias{gen_comb_char}
\title{Generate char vector of variable labels with dataset}
\usage{
gen_comb_char(datapath, dataset, language)
}
\description{
Generate char vector of variable labels with dataset
}
\keyword{internal}
