% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_labels_for_multi}
\alias{gen_labels_for_multi}
\title{Create variable labels for multi-variable selections}
\usage{
gen_labels_for_multi(datapath, dataset)
}
\description{
Create variable labels for multi-variable selections
}
\keyword{internal}
