% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_meta_var_table}
\alias{gen_meta_var_table}
\title{Generate metadata table for a single variable}
\usage{
gen_meta_var_table(datapath, dataset, sel_var, language, value_labels)
}
\description{
Generate metadata table for a single variable
}
\keyword{internal}
