% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_functions.R
\name{gen_parallel_spells_r}
\alias{gen_parallel_spells_r}
\title{function to generate educational qualification preparation. R.}
\usage{
gen_parallel_spells_r(format = "harmonized")
}
\description{
function to generate educational qualification preparation. R.
}
\keyword{internal}
