% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_functions.R
\name{gen_parallel_spells_stata}
\alias{gen_parallel_spells_stata}
\title{function to generate educational qualification preparation. Stata.}
\usage{
gen_parallel_spells_stata(format = "harmonized")
}
\description{
function to generate educational qualification preparation. Stata.
}
\keyword{internal}
