% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_functions.R
\name{gen_qualification_prep_code_r}
\alias{gen_qualification_prep_code_r}
\title{function to generate educational qualification preparation. R}
\usage{
gen_qualification_prep_code_r(english, SC, suf_version, suf_version_short)
}
\description{
function to generate educational qualification preparation. R
}
\keyword{internal}
