% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_functions.R
\name{gen_qualification_prep_code_stata}
\alias{gen_qualification_prep_code_stata}
\title{function to generate educational qualification preparation. Stata.}
\usage{
gen_qualification_prep_code_stata(english, sc, suf_version, suf_version_short)
}
\description{
function to generate educational qualification preparation. Stata.
}
\keyword{internal}
