% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_questiontext}
\alias{gen_questiontext}
\title{Generate question text for a variable}
\usage{
gen_questiontext(datapath, dataset, col_select, variable, language)
}
\description{
Generate question text for a variable
}
\keyword{internal}
