% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_script.R
\name{gen_script}
\alias{gen_script}
\title{function that generates the scripts}
\usage{
gen_script(
  datapath_conv,
  datapath_local,
  suf_version,
  suf_version_short,
  dataformat,
  subformat,
  datalist,
  prio,
  english,
  set_missings,
  parallel,
  further_training,
  education,
  children
)
}
\description{
function that generates the scripts
}
\keyword{internal}
