% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_functions.R
\name{generate_strings}
\alias{generate_strings}
\title{function to generate R strings that read and merge neps datafiles to the person year dataset based on a list that is generated by the app user selecting variables in different datasets in the data transformation tab}
\usage{
generate_strings(data_list, suf_version, english)
}
\description{
function to generate R strings that read and merge neps datafiles to the person year dataset based on a list that is generated by the app user selecting variables in different datasets in the data transformation tab
}
\keyword{internal}
