% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getWindowsDrives}
\alias{getWindowsDrives}
\title{# function that returns string about starting cohort in form: 'sc3':'sc6'}
\usage{
getWindowsDrives()
}
\description{
# function that returns string about starting cohort in form: 'sc3':'sc6'
}
\keyword{internal}
