% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identify_sc}
\alias{identify_sc}
\title{inspects filenames in a dir and return sc cohort identifier}
\usage{
identify_sc(datapath)
}
\description{
inspects filenames in a dir and return sc cohort identifier
}
\keyword{internal}
