% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_string_to_position}
\alias{move_string_to_position}
\title{helper function to change order of strings in vector}
\usage{
move_string_to_position(vec, string_to_move, target_position)
}
\description{
helper function to change order of strings in vector
}
\keyword{internal}
