% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_variable_to_position}
\alias{move_variable_to_position}
\title{helper function to change order of vars in dataframe}
\usage{
move_variable_to_position(df, var_to_move, after_var)
}
\description{
helper function to change order of vars in dataframe
}
\keyword{internal}
