% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEPScribe-package.R
\docType{package}
\name{NEPScribe-package}
\alias{NEPScribe}
\alias{NEPScribe-package}
\title{NEPScribe: Explore NEPS Datasets and Generate Data Preparation Scripts}
\description{
Provides a Shiny application to explore NEPS Scientific Use File datasets and to generate Stata and R scripts that transform them into a person-year format.
}
\author{
\strong{Maintainer}: Alexander Helbig \email{alexander.helbig@wzb.eu}

}
\keyword{internal}
