% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_meta_for_input_list}
\alias{process_meta_for_input_list}
\title{Process metadata for input selection list}
\usage{
process_meta_for_input_list(datapath, dataset)
}
\description{
Process metadata for input selection list
}
\keyword{internal}
