% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_dta_eng_labels}
\alias{read_dta_eng_labels}
\title{Read Stata dataset with English labels assigned}
\usage{
read_dta_eng_labels(datasetpath, col_select = NULL)
}
\description{
Read Stata dataset with English labels assigned
}
\keyword{internal}
