% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_exp_fields}
\alias{read_exp_fields}
\title{Read expansion fields from a dataset}
\usage{
read_exp_fields(datapath, cols = NULL, attr_type = NULL, only_value = FALSE)
}
\description{
Read expansion fields from a dataset
}
\keyword{internal}
