% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readstata13_label}
\alias{readstata13_label}
\title{function to read dta and assign variable labels (which are attracted but not yet assigned) to the variables}
\usage{
readstata13_label(x)
}
\description{
function to read dta and assign variable labels (which are attracted but not yet assigned) to the variables
}
\keyword{internal}
