% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_prefix_suffix_capitalize_df}
\alias{remove_prefix_suffix_capitalize_df}
\title{helper function to remove neps prefix from variable names in df and capitalize first letter}
\usage{
remove_prefix_suffix_capitalize_df(df, colnames = TRUE)
}
\description{
helper function to remove neps prefix from variable names in df and capitalize first letter
}
\keyword{internal}
