% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_prefix_suffix_capitalize_vec}
\alias{remove_prefix_suffix_capitalize_vec}
\title{helper function to remove neps prefix from variable names in vector and capitalize first letter}
\usage{
remove_prefix_suffix_capitalize_vec(var_names)
}
\description{
helper function to remove neps prefix from variable names in vector and capitalize first letter
}
\keyword{internal}
