% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings.R
\name{settings_server}
\alias{settings_server}
\title{Settings Server Module}
\usage{
settings_server(id)
}
\description{
Provides a reactive list of global settings selected in the UI.
}
\keyword{internal}
