% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings.R
\name{settings_ui}
\alias{settings_ui}
\title{Settings UI Module}
\usage{
settings_ui(id)
}
\description{
Provides global settings for the app, including sidebar width and label language.
}
\keyword{internal}
