% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_all_to_factors}
\alias{transform_all_to_factors}
\title{function to check and transform variables to a factor if they are not of type factor or labelled doubles}
\usage{
transform_all_to_factors(data)
}
\description{
function to check and transform variables to a factor if they are not of type factor or labelled doubles
}
\keyword{internal}
