% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truncate_strings}
\alias{truncate_strings}
\title{function to truncate strings to 15 chars and return as unnamed char vec}
\usage{
truncate_strings(input_strings)
}
\description{
function to truncate strings to 15 chars and return as unnamed char vec
}
\keyword{internal}
