% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_exploration.R
\name{dataset_ui}
\alias{dataset_ui}
\title{Data Exploration Module}
\usage{
dataset_ui(id)
}
\description{
Internal Shiny module for exploring datasets.
}
\details{
This module provides UI and server components for selecting datasets, viewing dataset summaries.
It uses reactive programming to handle multiple datasets and cohort paths.
}
\keyword{internal}
