% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEPScribe-package.R
\docType{package}
\name{NEPScribe-package}
\alias{NEPScribe}
\alias{NEPScribe-package}
\title{NEPScribe: What the Package Does (Title Case)}
\description{
Explore NEPS datasets and generate scripts to transform the data.
}
\author{
\strong{Maintainer}: Alexander Helbig \email{alexander.helbig@wzb.eu}

}
\keyword{internal}
